// src/types/birthData.ts
export interface BirthData {
  birthDate: string;
  birthTime: string;
  birthPlace: string;
  city: string;
  country: string;
  countryCode: string;
  latitude: number;
  longitude: number;
  timezone: string;
  placeId: string;
}

export interface BirthDataFormData {
  birthDate: string;
  birthTime: string;
  birthPlace: string;
}